rename definitions to omniAccount
align with our preferred name of OmniAccount
