+++ tee-worker/identity/client-api/parachain-api/prepare-build/interfaces/definitions.ts
@@ -1,5 +1,5 @@
 export { default as identity } from "./identity/definitions";
 export { default as vc } from "./vc/definitions";
-export { default as omniaccount } from "./omniaccount/definitions";
+export { default as omniAccount } from "./omniAccount/definitions";
 export { default as trusted_operations } from "./trusted_operations/definitions";
 export { default as sidechain } from "./sidechain/definitions";
